Crate pallet_proxy

Proxy Module

A module allowing accounts to give permission to other accounts to dispatch types of calls from their signed origin.

The accounts to which permission is delegated may be requied to announce the action that they wish to execute some duration prior to execution happens. In this case, the target account may reject the announcement and in doing so, veto the execution.

proxy::Trait

Call

Overview

Interface

Dispatchable Functions

Re-exports

pub use weights::WeightInfo;

Modules

migration
Migration utilities for upgrading the Proxy pallet between its different versions.

weights
Weights for pallet_proxy THIS FILE WAS AUTO-GENERATED USING THE SUBSTRATE BENCHMARK CLI VERSION 2.0.0 DATE: 2020-10-27, STEPS: [50, ], REPEAT: 20, LOW RANGE: [], HIGH RANGE: [] EXECUTION: Some(Wasm), WASM-EXECUTION: Compiled, CHAIN: Some("dev"), DB CACHE: 128

Structs

Announcement
Details surrounding a specific instance of an announcement to make a call.

Announcements
The announcements made by the proxy (key).

Module

Proxies
The set of account proxies. Maps the account which has delegated to the accounts which are being delegated to, together with the amount held on deposit.

ProxyDefinition
The parameters under which a particular account has a proxy relationship with some other account.

Enums

Call
Dispatchable calls.

Error

RawEvent
Events for this module.

Traits

Trait
Configuration trait.

_
Provides version information about a pallet.

Type Definitions

Event
RawEvent specialized for the configuration Trait