Function pallet_proxy::pallet::dispatchables::create_pure

pub fn create_pure<T: Config>(
    proxy_type: T::ProxyType,
    delay: T::BlockNumber,
    index: u16
)

Spawn a fresh new account that is guaranteed to be otherwise inaccessible, and initialize it with a proxy of proxy_type for origin sender.

Requires a Signed origin.

proxy_type: The type of the proxy that the sender will be registered as over the new account. This will almost always be the most permissive ProxyType possible to allow for maximum flexibility.

index: A disambiguation index, in case this is called multiple times in the same transaction (e.g. with utility::batch). Unless you’re using batch you probably just want to use 0.

delay: The announcement period required of the initial proxy. Will generally be zero.

Fails with Duplicate if this has already been called in this transaction, from the same sender, with the same parameters.

Fails if there are insufficient funds to pay for deposit.

NOTE: This function is an automatically generated, doc only, uncallable stub. designed to document the create_pure variant of the pallet Call enum. You should not attempt to call this function directly.