pallet_proxy::pallet::dispatchables

Function reject_announcement

pub fn reject_announcement<T: Config>(
    delegate: <<T as Config>::Lookup as StaticLookup>::Source,
    call_hash: <<T as Config>::CallHasher as Hash>::Output,
)

Remove the given announcement of a delegate.

May be called by a target (proxied) account to remove a call that one of their delegates (delegate) has announced they want to execute. The deposit is returned.

The dispatch origin for this call must be Signed.

Parameters:

delegate: The account that previously announced the call.

call_hash: The hash of the call to be made.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::reject_announcement.