Type Alias pallet_proxy::pallet::Module

pub type Module<T> = Pallet<T>;

👎Deprecated: use Pallet instead

Type alias to Pallet, to be used by construct_runtime.

Generated by pallet attribute macro.

Aliased Type

struct Module<T>(/* private fields */);

Implementations

impl<T: Config> Pallet<T>

pub fn proxy( origin: OriginFor<T>, real: <<T as Config>::Lookup as StaticLookup>::Source, force_proxy_type: Option<T::ProxyType>, call: Box<<T as Config>::RuntimeCall> ) -> DispatchResult

Dispatch the given call from an account that the sender is authorised for through add_proxy.

The dispatch origin for this call must be Signed.

Parameters:

real: The account that the proxy will make a call on behalf of.

force_proxy_type: Specify the exact proxy type to be used and checked for this call.

call: The call to be made by the real account.

pub fn add_proxy( origin: OriginFor<T>, delegate: <<T as Config>::Lookup as StaticLookup>::Source, proxy_type: T::ProxyType, delay: BlockNumberFor<T> ) -> DispatchResult

Register a proxy account for the sender that is able to make calls on its behalf.

The dispatch origin for this call must be Signed.

Parameters:

proxy: The account that the caller would like to make a proxy.

proxy_type: The permissions allowed for this proxy account.

delay: The announcement period required of the initial proxy. Will generally be zero.

pub fn remove_proxy( origin: OriginFor<T>, delegate: <<T as Config>::Lookup as StaticLookup>::Source, proxy_type: T::ProxyType, delay: BlockNumberFor<T> ) -> DispatchResult

Unregister a proxy account for the sender.

The dispatch origin for this call must be Signed.

Parameters:

proxy: The account that the caller would like to remove as a proxy.

proxy_type: The permissions currently enabled for the removed proxy account.

pub fn remove_proxies(origin: OriginFor<T>) -> DispatchResult

Unregister all proxy accounts for the sender.

The dispatch origin for this call must be Signed.

WARNING: This may be called on accounts created by pure, however if done, then the unreserved fees will be inaccessible. All access to this account will be lost.

pub fn create_pure( origin: OriginFor<T>, proxy_type: T::ProxyType, delay: BlockNumberFor<T>, index: u16 ) -> DispatchResult

Spawn a fresh new account that is guaranteed to be otherwise inaccessible, and initialize it with a proxy of proxy_type for origin sender.

Requires a Signed origin.

proxy_type: The type of the proxy that the sender will be registered as over the new account. This will almost always be the most permissive ProxyType possible to allow for maximum flexibility.

index: A disambiguation index, in case this is called multiple times in the same transaction (e.g. with utility::batch). Unless you’re using batch you probably just want to use 0.

delay: The announcement period required of the initial proxy. Will generally be zero.

Fails with Duplicate if this has already been called in this transaction, from the same sender, with the same parameters.

Fails if there are insufficient funds to pay for deposit.

pub fn kill_pure( origin: OriginFor<T>, spawner: <<T as Config>::Lookup as StaticLookup>::Source, proxy_type: T::ProxyType, index: u16, height: BlockNumberFor<T>, ext_index: u32 ) -> DispatchResult

Removes a previously spawned pure proxy.

WARNING: All access to this account will be lost. Any funds held in it will be inaccessible.

Requires a Signed origin, and the sender account must have been created by a call to pure with corresponding parameters.

spawner: The account that originally called pure to create this account.

index: The disambiguation index originally passed to pure. Probably 0.

proxy_type: The proxy type originally passed to pure.

height: The height of the chain when the call to pure was processed.

ext_index: The extrinsic index in which the call to pure was processed.

Fails with NoPermission in case the caller is not a previously created pure account whose pure call has corresponding parameters.

pub fn announce( origin: OriginFor<T>, real: <<T as Config>::Lookup as StaticLookup>::Source, call_hash: <<T as Config>::CallHasher as Hash>::Output ) -> DispatchResult

Publish the hash of a proxy-call that will be made in the future.

This must be called some number of blocks before the corresponding proxy is attempted if the delay associated with the proxy relationship is greater than zero.

No more than MaxPending announcements may be made at any one time.

This will take a deposit of AnnouncementDepositFactor as well as AnnouncementDepositBase if there are no other pending announcements.

The dispatch origin for this call must be Signed and a proxy of real.

Parameters:

real: The account that the proxy will make a call on behalf of.

call_hash: The hash of the call to be made by the real account.

pub fn remove_announcement( origin: OriginFor<T>, real: <<T as Config>::Lookup as StaticLookup>::Source, call_hash: <<T as Config>::CallHasher as Hash>::Output ) -> DispatchResult

Remove a given announcement.

May be called by a proxy account to remove a call they previously announced and return the deposit.

The dispatch origin for this call must be Signed.

Parameters:

real: The account that the proxy will make a call on behalf of.

call_hash: The hash of the call to be made by the real account.

pub fn reject_announcement( origin: OriginFor<T>, delegate: <<T as Config>::Lookup as StaticLookup>::Source, call_hash: <<T as Config>::CallHasher as Hash>::Output ) -> DispatchResult

Remove the given announcement of a delegate.

May be called by a target (proxied) account to remove a call that one of their delegates (delegate) has announced they want to execute. The deposit is returned.

The dispatch origin for this call must be Signed.

Parameters:

delegate: The account that previously announced the call.

call_hash: The hash of the call to be made.

pub fn proxy_announced( origin: OriginFor<T>, delegate: <<T as Config>::Lookup as StaticLookup>::Source, real: <<T as Config>::Lookup as StaticLookup>::Source, force_proxy_type: Option<T::ProxyType>, call: Box<<T as Config>::RuntimeCall> ) -> DispatchResult

Dispatch the given call from an account that the sender is authorized for through add_proxy.

Removes any corresponding announcement(s).

The dispatch origin for this call must be Signed.

Parameters:

real: The account that the proxy will make a call on behalf of.

force_proxy_type: Specify the exact proxy type to be used and checked for this call.

call: The call to be made by the real account.

impl<T: Config> Pallet<T>

pub fn proxies<KArg>( k: KArg ) -> (BoundedVec<ProxyDefinition<T::AccountId, T::ProxyType, BlockNumberFor<T>>, T::MaxProxies>, <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance)where KArg: EncodeLike<T::AccountId>,

An auto-generated getter for Proxies.

impl<T: Config> Pallet<T>

pub fn announcements<KArg>( k: KArg ) -> (BoundedVec<Announcement<T::AccountId, <<T as Config>::CallHasher as Hash>::Output, BlockNumberFor<T>>, T::MaxPending>, <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance)where KArg: EncodeLike<T::AccountId>,

An auto-generated getter for Announcements.

impl<T: Config> Pallet<T>

pub fn pure_account( who: &T::AccountId, proxy_type: &T::ProxyType, index: u16, maybe_when: Option<(BlockNumberFor<T>, u32)> ) -> T::AccountId

Calculate the address of an pure account.

who: The spawner account.

proxy_type: The type of the proxy that the sender will be registered as over the new account. This will almost always be the most permissive ProxyType possible to allow for maximum flexibility.

index: A disambiguation index, in case this is called multiple times in the same transaction (e.g. with utility::batch). Unless you’re using batch you probably just want to use 0.

maybe_when: The block height and extrinsic index of when the pure account was created. None to use current block height and extrinsic index.

pub fn add_proxy_delegate( delegator: &T::AccountId, delegatee: T::AccountId, proxy_type: T::ProxyType, delay: BlockNumberFor<T> ) -> DispatchResult

Register a proxy account for the delegator that is able to make calls on its behalf.

Parameters:

delegator: The delegator account.

delegatee: The account that the delegator would like to make a proxy.

proxy_type: The permissions allowed for this proxy account.

delay: The announcement period required of the initial proxy. Will generally be zero.

pub fn remove_proxy_delegate( delegator: &T::AccountId, delegatee: T::AccountId, proxy_type: T::ProxyType, delay: BlockNumberFor<T> ) -> DispatchResult

Unregister a proxy account for the delegator.

Parameters:

delegator: The delegator account.

delegatee: The account that the delegator would like to make a proxy.

proxy_type: The permissions allowed for this proxy account.

delay: The announcement period required of the initial proxy. Will generally be zero.

pub fn deposit( num_proxies: u32 ) -> <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance

pub fn find_proxy( real: &T::AccountId, delegate: &T::AccountId, force_proxy_type: Option<T::ProxyType> ) -> Result<ProxyDefinition<T::AccountId, T::ProxyType, BlockNumberFor<T>>, DispatchError>

pub fn remove_all_proxy_delegates(delegator: &T::AccountId)

Removes all proxy delegates for a given delegator.

Parameters:

delegator: The delegator account.