Type Definition pallet_proxy::pallet::Proxies

type Proxies<T: Config> = StorageMap<_GeneratedPrefixForStorageProxies<T>, Twox64Concat, T::AccountId, (BoundedVec<ProxyDefinition<T::AccountId, T::ProxyType, T::BlockNumber>, T::MaxProxies>, <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance), ValueQuery>;

The set of account proxies. Maps the account which has delegated to the accounts which are being delegated to, together with the amount held on deposit.