Trait pallet_proxy::Trait

pub trait Trait: Trait {
    type Event: From<Event<Self>> + Into<Self::Event>;
    type Call: Parameter + Dispatchable<Origin = Self::Origin, PostInfo = PostDispatchInfo> + GetDispatchInfo + From<Call<Self>> + IsSubType<Call<Self>> + IsType<Self::Call>;
    type Currency: ReservableCurrency<Self::AccountId>;
    type ProxyType: Parameter + Member + Ord + PartialOrd + InstanceFilter<Self::Call> + Default;
    type ProxyDepositBase: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>;
    type ProxyDepositFactor: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>;
    type MaxProxies: Get<u16>;
    type WeightInfo: WeightInfo;
}

Configuration trait.

Associated Types

type Event: From<Event<Self>> + Into<Self::Event>

The overarching event type.

type Call: Parameter + Dispatchable<Origin = Self::Origin, PostInfo = PostDispatchInfo> + GetDispatchInfo + From<Call<Self>> + IsSubType<Call<Self>> + IsType<Self::Call>

The overarching call type.

type Currency: ReservableCurrency<Self::AccountId>

The currency mechanism.

type ProxyType: Parameter + Member + Ord + PartialOrd + InstanceFilter<Self::Call> + Default

A kind of proxy; specified with the proxy and passed in to the IsProxyable fitler. The instance filter determines whether a given call may be proxied under this type.

IMPORTANT: Default must be provided and MUST BE the the most permissive value.

type ProxyDepositBase: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>

The base amount of currency needed to reserve for creating a proxy.

This is held for an additional storage item whose value size is sizeof(Balance) bytes and whose key size is sizeof(AccountId) bytes.

type ProxyDepositFactor: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>

The amount of currency needed per proxy added.

This is held for adding 32 bytes plus an instance of ProxyType more into a pre-existing storage value.

type MaxProxies: Get<u16>

The maximum amount of proxies allowed for a single account.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

Loading content...