Crate pallet_proxy

Proxy Pallet

A pallet allowing accounts to give permission to other accounts to dispatch types of calls from their signed origin.

The accounts to which permission is delegated may be required to announce the action that they wish to execute some duration prior to execution happens. In this case, the target account may reject the announcement and in doing so, veto the execution.

Config

Call

Re-exports

pub use weights::WeightInfo;

pub use pallet::*;

Modules

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

weights

Autogenerated weights for pallet_proxy

Structs

Announcement

Details surrounding a specific instance of an announcement to make a call.

ProxyDefinition

The parameters under which a particular account has a proxy relationship with some other account.