Crate pallet_proxy

Proxy Module

A module allowing accounts to give permission to other accounts to dispatch types of calls from their signed origin.

proxy::Trait

Call

Overview

Interface

Dispatchable Functions

Structs

Module

Proxies
The set of account proxies. Maps the account which has delegated to the accounts which are being delegated to, together with the amount held on deposit.

Enums

Call
Dispatchable calls.

Error

RawEvent
Events for this module.

Traits

Trait
Configuration trait.

WeightInfo

Type Definitions

Event
RawEvent specialized for the configuration Trait