Trait pallet_proxy::pallet::Config

pub trait Config: Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type Call: Parameter + Dispatchable<Origin = Self::Origin> + GetDispatchInfo + From<Call<Self>> + IsSubType<Call<Self>> + IsType<<Self as Config>::Call>;
    type Currency: ReservableCurrency<Self::AccountId>;
    type ProxyType: Parameter + Member + Ord + PartialOrd + InstanceFilter<<Self as Config>::Call> + Default + MaxEncodedLen;
    type ProxyDepositBase: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>;
    type ProxyDepositFactor: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>;
    type MaxProxies: Get<u32>;
    type WeightInfo: WeightInfo;
    type MaxPending: Get<u32>;
    type CallHasher: Hash;
    type AnnouncementDepositBase: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>;
    type AnnouncementDepositFactor: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>;
}

Configuration trait.

Required Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

The overarching event type.

type Call: Parameter + Dispatchable<Origin = Self::Origin> + GetDispatchInfo + From<Call<Self>> + IsSubType<Call<Self>> + IsType<<Self as Config>::Call>

The overarching call type.

type Currency: ReservableCurrency<Self::AccountId>

The currency mechanism.

type ProxyType: Parameter + Member + Ord + PartialOrd + InstanceFilter<<Self as Config>::Call> + Default + MaxEncodedLen

A kind of proxy; specified with the proxy and passed in to the IsProxyable fitler. The instance filter determines whether a given call may be proxied under this type.

IMPORTANT: Default must be provided and MUST BE the the most permissive value.

type ProxyDepositBase: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>

The base amount of currency needed to reserve for creating a proxy.

This is held for an additional storage item whose value size is sizeof(Balance) bytes and whose key size is sizeof(AccountId) bytes.

type ProxyDepositFactor: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>

The amount of currency needed per proxy added.

This is held for adding 32 bytes plus an instance of ProxyType more into a pre-existing storage value. Thus, when configuring ProxyDepositFactor one should take into account 32 + proxy_type.encode().len() bytes of data.

type MaxProxies: Get<u32>

The maximum amount of proxies allowed for a single account.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type MaxPending: Get<u32>

The maximum amount of time-delayed announcements that are allowed to be pending.

type CallHasher: Hash

The type of hash used for hashing the call.

type AnnouncementDepositBase: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>

The base amount of currency needed to reserve for creating an announcement.

This is held when a new storage item holding a Balance is created (typically 16 bytes).

type AnnouncementDepositFactor: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>

The amount of currency needed per announcement made.

This is held for adding an AccountId, Hash and BlockNumber (typically 68 bytes) into a pre-existing storage value.